Attempt at implementing a rate-limit sink
Added two files modeled after other sinks in attempt to create a rate-limit sink.
Updated all of the rate code to use an integer rate rather than a double rate.
Not sure how to calculate rate or how I would make the call to write to the log.
Attempt to make changes discussed in #108.