Changes to the rate-limit sink
Made more of the changes suggested, including how to calculate rate.
Modfied the clean-up to be more effective (as suggested).
Attempted to calculate the rate in the rate-limit-sink class using timestamps.
Chaged 'write' return value to be the return value of writing to the other sink.
Added a 'sink_owned' variable.

Need to work on: getting the timespec values (especially the first one).